+++ checkpoint/CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Ignore not-exists and not-dir errors while building step metadata in
  _StandardNameFormat.
 
+### Changed
+- Return `StepMetadata` from `CompositeCheckpointHandler.metadata()`.
+
 ## [0.10.2] - 2024-12-04
 
 ### Added